+++ util/logging/kafkaconfigurationmessage_test.go
@@ -5,11 +5,12 @@ import (
 )
 
 func TestReportKafkaConfiguration_Zero(t *testing.T) {
+	var hook *liveKafkaSink
+	var cfg Config
+
+	reportKafkaConfig(hook, cfg, SilentLogSet())
 	AssertReportFields(t,
 		func(ls LogSink) {
-			var hook *kafkaSink
-			var cfg Config
-
 			reportKafkaConfig(hook, cfg, ls)
 		},
 		StandardVariableFields,
@@ -18,13 +19,16 @@ func TestReportKafkaConfiguration_Zero(t *testing.T) {
 			"severity":                   WarningLevel,
 			"call-stack-message":         "Not connecting to Kafka.",
 			"sous-successful-connection": false,
+			"kafka-logging-topic":        "",
+			"kafka-brokers":              "",
+			"kafka-logger-id":            "",
 		})
 }
 
 func TestReportKafkaConfiguration_Complete(t *testing.T) {
 	AssertReportFields(t,
 		func(ls LogSink) {
-			hook := &kafkaSink{}
+			hook := &liveKafkaSink{}
 			cfg := Config{}
 			cfg.Kafka.Topic = "test-topic"
 			cfg.Kafka.BrokerList = "broker1,broker2,broker3"
@@ -38,7 +42,6 @@ func TestReportKafkaConfiguration_Complete(t *testing.T) {
 			"call-stack-message":         "Connecting to Kafka",
 			"kafka-logging-topic":        "test-topic",
 			"kafka-brokers":              "broker1,broker2,broker3",
-			"kafka-logging-levels":       "CriticalLevel",
 			"kafka-logger-id":            "",
 			"sous-successful-connection": true,
 		})